% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBK_eq.R
\name{calcBK_eq}
\alias{calcBK_eq}
\title{CalcBK_eq}
\usage{
calcBK_eq(mergedData = NULL)
}
\arguments{
\item{mergedData}{The output of the getAndFormatData function which contains the necessary formatted data. Columns
1-12 include variables obtained directly from the water quality data product (DP1.20288.001). Descriptions and units
of these variables can be found in the readme_20288 file.  Columns 13-14 include variables obtained directly from
the barometric pressure data product (DP1.00004.001). Descriptions and units of these variables can be found in the
readme_00004 file. Column 15 "atm" is the barometric pressure converted to units of atmospheres.  Columns 16-17 1
min interpolated values derived from the 5 min values found in the water temperature data product (DP1.20053.001).
Descriptions and units of these variables can be found in the readme_20053 file.   [dataFrame]}
}
\value{
This script returns the percent DO saturation (units %) corrected to local atmospheric pressure, and a
quality flag which summarizes the quality flags of the data products used in the calculation.
}
\description{
This R script calculates the dissolved oxygen (DO) percent saturation at local atmospheric pressure
from the DO percent saturation reported in the water quality data product (DP1.20288.001), which is
referenced to sea level.  This is done using the Benson-Krause pressure correction factor found in Standard Methods
(2005) and used by the USGS since 2011.

The pressure correction factor Fp is given as:

Fp = (P-u)(1-zP) / (1-u)(1-z)

where P is the barometric pressure in atmospheres, u is the vapor pressure of water in atmospheres, and z is related
to the second virial coefficient of oxygen.

u = exp(11.8571 - (3840.7/(T+273.15)) - (216961/(T+273.15)^2))

and

z = 0.000975 - 0.00001426T + 0.00000006436T^2

where T is water temperature in degrees C.

Pressure is obtained from the barometric pressure data product (DP1.00004.001), and water temperature is obtained
from the water temperature data product (DP1.20053.001).
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Bobby Hensley \email{hensley@battelleecology.org} \cr
}
