% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndFormatData.R
\name{getAndFormatData}
\alias{getAndFormatData}
\title{getAndFormatData}
\usage{
getAndFormatData(siteName = NULL, startDate = NULL, endDate = NULL)
}
\arguments{
\item{siteName}{The 4 digit NEON site code, e.g. "LEWI" [character]}

\item{startDate}{The data start date foratted as "YYYY-MM" [character]}

\item{endDate}{The data end date foratted as "YYYY-MM" [character]}
}
\value{
This script produces table of properly formatted data.
}
\description{
This R script that downloads and formats the data necessary to correct dissolved oxygen (DO) percent
saturation to local atmospheric pressure from the DO percent saturation reported in the water quality data product
(DP1.20288.001), which is referenced to sea level.  This can be done using either the Benson-Krause
equation used by the USGS, or a simplified version found in the YSI EXO2 manual.

The relevant data products are:

water quality (DP1.20288.001)
barometric pressure (DP1.00004.001)
water temperature (DP1.20053.001) - Note water temp is only used for Benson-Krause method.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Bobby Hensley \email{hensley@battelleecology.org} \cr
}
